+++ CHANGELOG.md
@@ -1,32 +1,58 @@
-# 1.0.1 
+## 1.1.2
 
-## Chore
+### Added Features
+
+- add Electron platform
+
+### Bug fixes
+
+- fix no Listeners for Electron platform
+
+## 1.1.1
+
+### Bug fixes
+
+- test Electron on Windows OS
+
+## 1.1.0 
+
+### Chore
+
+- update to @capacitor/core@3.2.0
+
+### Added Features
+
+- add Electron platform
+
+## 1.0.1 
+
+### Chore
 
 - update to @capacitor/core@3.1.2
 
-## Bug fixes
+### Bug fixes
 
 - Tag SUSPENSE has been removed
 - the initialization of the database has been moved from Home to main.js
 
-# 1.0.0
+## 1.0.0
 
-## Chore
+### Chore
 
 - update to @capacitor/core@3.1.1
 
-# 0.3.0
+## 0.3.0
 
-## Chore
+### Chore
 
 - update to @capacitor/core@2.4.5
 
-# 0.2.0
+## 0.2.0
 
-## Chore
+### Chore
 
 - update to @capacitor/core@2.4.2
 
-# 0.1.0
+## 0.1.0
 
 - Initial commit